Update dependencies and GitHub action versions
This commit includes updates to the versions of dependencies in the pom.xml file. In addition, the versions of actions used in the Qodana code quality GitHub workflow file have been updated. This ensures our project is being built and tested against the latest plugin and action versions for improved performance and security.